+++ worker/src/scheduled.ts
@@ -6,31 +6,48 @@ import { CleanupSettings } from './models';
 
 export async function scheduled(event: ScheduledEvent, env: Bindings, ctx: any) {
     console.log("Scheduled event: ", event);
-    const value = await getJsonSetting(
+    const autoCleanupSetting = await getJsonSetting<CleanupSettings>(
         { env: env, } as Context<HonoCustomType>,
         CONSTANTS.AUTO_CLEANUP_KEY
     );
-    const autoCleanupSetting = new CleanupSettings(value);
+    if (!autoCleanupSetting) {
+        console.log("No auto cleanup settings found, skipping cleanup.");
+        return;
+    }
     console.log("autoCleanupSetting:", JSON.stringify(autoCleanupSetting));
-    if (autoCleanupSetting.enableMailsAutoCleanup && autoCleanupSetting.cleanMailsDays > 0) {
+    if (autoCleanupSetting.enableMailsAutoCleanup) {
         await cleanup(
             { env: env, } as Context<HonoCustomType>,
             "mails",
             autoCleanupSetting.cleanMailsDays
         );
     }
-    if (autoCleanupSetting.enableUnknowMailsAutoCleanup && autoCleanupSetting.cleanUnknowMailsDays > 0) {
+    if (autoCleanupSetting.enableUnknowMailsAutoCleanup) {
         await cleanup(
             { env: env, } as Context<HonoCustomType>,
             "mails_unknow",
             autoCleanupSetting.cleanUnknowMailsDays
         );
     }
-    if (autoCleanupSetting.enableSendBoxAutoCleanup && autoCleanupSetting.cleanSendBoxDays > 0) {
+    if (autoCleanupSetting.enableSendBoxAutoCleanup) {
         await cleanup(
             { env: env, } as Context<HonoCustomType>,
             "sendbox",
             autoCleanupSetting.cleanSendBoxDays
+        );
+    }
+    if (autoCleanupSetting.enableInactiveAddressAutoCleanup) {
+        await cleanup(
+            { env: env, } as Context<HonoCustomType>,
+            "inactiveAddress",
+            autoCleanupSetting.cleanInactiveAddressDays
+        );
+    }
+    if (autoCleanupSetting.enableAddressAutoCleanup) {
+        await cleanup(
+            { env: env, } as Context<HonoCustomType>,
+            "addressCreated",
+            autoCleanupSetting.cleanAddressDays
         );
     }
 }