+++ worker/src/types.d.ts
@@ -16,6 +16,7 @@ type Bindings = {
     DEFAULT_LANG: string | undefined
     TITLE: string | undefined
     ANNOUNCEMENT: string | undefined | null
+    ALWAYS_SHOW_ANNOUNCEMENT: string | boolean | undefined
     PREFIX: string | undefined
     ADDRESS_CHECK_REGEX: string | undefined
     ADDRESS_REGEX: string | undefined
@@ -23,6 +24,8 @@ type Bindings = {
     MAX_ADDRESS_LEN: string | number | undefined
     DEFAULT_DOMAINS: string | string[] | undefined
     DOMAINS: string | string[] | undefined
+    DISABLE_CUSTOM_ADDRESS_NAME: string | boolean | undefined
+    CREATE_ADDRESS_DEFAULT_DOMAIN_FIRST: string | boolean | undefined
     ADMIN_USER_ROLE: string | undefined
     USER_DEFAULT_ROLE: string | UserRole | undefined
     USER_ROLES: string | UserRole[] | undefined